연산될 갯수체크

+++ dotori.go
@@ -42,7 +42,7 @@ var (
 	flagHTTPPort        = flag.String("http", "", "Web Service Port Number")
 	flagPagenum         = flag.Int64("pagenum", 9, "maximum number of items in a page")
 	flagCookieAge       = flag.Int("cookieage", 4, "cookie age (hour)") // MPAA 기준 4시간이다.
-	flagProcessNum      = flag.Int("processnum", 1, "number of process")
+	flagProcessNum      = flag.Int64("processnum", 1, "number of process")
 	flagProcessInterval = flag.Int("processinterval", 10, "seconds of interval between processes") // 기본 10초
 	flagCertFullchanin  = flag.String("certfullchanin", "", "certification fullchain path")
 	flagCertPrivkey     = flag.String("certprivkey", "", "certification privkey path")
